requirements.txt
python-telegram-bot[webhooks]==21.6
python-dotenv==1.0.1
docker==7.1.0
aiosqlite==0.20.0
pyyaml==6.0.2
httpx==0.28.1
